Move redirect logic to middleware and update tests
