+++ client/my-sites/checkout/controller.jsx
@@ -1,9 +1,4 @@
-import {
-	isJetpackLegacyItem,
-	isJetpackLegacyTermUpgrade,
-	isJetpackPlanSlug,
-	isJetpackProductSlug,
-} from '@automattic/calypso-products';
+import { isJetpackLegacyItem, isJetpackLegacyTermUpgrade } from '@automattic/calypso-products';
 import page from '@automattic/calypso-router';
 import debugFactory from 'debug';
 import { useTranslate } from 'i18n-calypso';
@@ -50,11 +45,7 @@ import UpsellNudge, {
 	CONCIERGE_QUICKSTART_SESSION,
 	PROFESSIONAL_EMAIL_UPSELL,
 } from './upsell-nudge';
-import {
-	getProductSlugFromContext,
-	isContextJetpackSitelessCheckout,
-	isContextSourceMyJetpack,
-} from './utils';
+import { getProductSlugFromContext, isContextJetpackSitelessCheckout } from './utils';
 
 const debug = debugFactory( 'calypso:checkout-controller' );
 
@@ -167,10 +158,6 @@ export function checkout( context, next ) {
 			return true;
 		}
 
-		if ( isContextSourceMyJetpack( context ) ) {
-			return true;
-		}
-
 		if ( isGiftPurchase ) {
 			return true;
 		}
@@ -188,19 +175,6 @@ export function checkout( context, next ) {
 	}
 
 	const product = getProductSlugFromContext( context );
-
-	if ( ( isJetpackPlanSlug( product ) || isJetpackProductSlug( product ) ) && isLoggedOut ) {
-		// Redirect to the siteless checkout page
-		const redirectUrl = addQueryArgs(
-			{
-				connect_after_checkout: true,
-				from_site_slug: context.query.site,
-				admin_url: context.query.redirect_to.split( '?' )[ 0 ],
-			},
-			context.path.replace( /checkout\/[^?/]+\//, 'checkout/jetpack/' )
-		);
-		page( redirectUrl );
-	}
 
 	if ( 'thank-you' === product ) {
 		return;
