+++ client/my-sites/checkout/index.js
@@ -2,7 +2,7 @@ import page from '@automattic/calypso-router';
 import {
 	makeLayout,
 	redirectLoggedOut,
-	redirectLoggedOutIfNotFromMyJetpack,
+	redirectMyJetpack,
 	render as clientRender,
 	setLocaleMiddleware,
 } from 'calypso/controller';
@@ -300,7 +300,8 @@ export default function () {
 	page(
 		`/checkout/:domainOrProduct`,
 		setLocaleMiddleware(),
-		redirectLoggedOutIfNotFromMyJetpack,
+		redirectMyJetpack,
+		redirectLoggedOut,
 		siteSelection,
 		redirectJetpackLegacyPlans,
 		checkout,
@@ -311,7 +312,8 @@ export default function () {
 	page(
 		`/checkout/:product/:domainOrProduct`,
 		setLocaleMiddleware(),
-		redirectLoggedOutIfNotFromMyJetpack,
+		redirectMyJetpack,
+		redirectLoggedOut,
 		siteSelection,
 		redirectJetpackLegacyPlans,
 		checkout,
