+++ client/my-sites/checkout/test/controller.js
@@ -9,21 +9,12 @@ import {
 } from '@automattic/calypso-products';
 import * as page from '@automattic/calypso-router';
 import configureStore from 'redux-mock-store';
-import addQueryArgs from 'calypso/lib/url/add-query-args';
 import { COMPARE_PLANS_QUERY_PARAM } from '../../plans/jetpack-plans/plan-upgrade/constants';
-import { checkout, redirectJetpackLegacyPlans } from '../controller';
+import { redirectJetpackLegacyPlans } from '../controller';
 import * as utils from '../utils';
 
 jest.mock( '@automattic/calypso-router' );
-jest.mock( '../utils', () => {
-	const originalModule = jest.requireActual( '../utils' );
-
-	return {
-		__esModule: true,
-		...originalModule,
-		getProductSlugFromContext: jest.fn(),
-	};
-} );
+jest.mock( '../utils' );
 
 const mockStore = configureStore();
 
@@ -74,78 +65,5 @@ describe( 'redirectJetpackLegacyPlans', () => {
 
 		expect( spy ).toHaveBeenCalledWith( redirectUrl );
 		expect( next ).not.toHaveBeenCalled();
-	} );
-} );
-
-describe( 'checkout', () => {
-	let pageSpy;
-	let next;
-
-	const query = {
-		site: 'example.com',
-		redirect_to: encodeURIComponent( 'http://example.com/wp-admin/admin.php?page=my-jetpack' ),
-		source: 'my-jetpack',
-	};
-
-	const context = {
-		store: mockStore( {
-			currentUser: {
-				id: null,
-			},
-			ui: {
-				selectedSiteId: null,
-			},
-		} ),
-		query,
-		params: {},
-		path: addQueryArgs( query, `/checkout/123123/jetpack_backup_t1_yearly` ),
-		pathname: '/checkout/123123/jetpack_backup_t1_yearly',
-	};
-
-	beforeEach( () => {
-		pageSpy = jest.spyOn( page, 'default' );
-		next = jest.fn();
-	} );
-
-	afterEach( () => {
-		pageSpy.mockRestore();
-		next.mockRestore();
-	} );
-
-	it( 'should redirect checkout with site id to siteless checkout if user is not logged in', () => {
-		const redirectUrl = addQueryArgs(
-			{
-				connect_after_checkout: true,
-				from_site_slug: context.query.site,
-				admin_url: context.query.redirect_to.split( '?' )[ 0 ],
-			},
-			context.path.replace( /checkout\/[^?/]+\//, 'checkout/jetpack/' )
-		);
-
-		checkout( context, next );
-
-		expect( pageSpy ).toHaveBeenCalledWith( redirectUrl );
-		expect( next ).toHaveBeenCalled();
-	} );
-
-	it( 'should redirect checkout with site slug to siteless checkout if user is not logged in', () => {
-		const contextWithSiteSlug = {
-			...context,
-			path: addQueryArgs( query, `/checkout/example.com/jetpack_backup_t1_yearly` ),
-			pathname: '/checkout/example.com/jetpack_backup_t1_yearly',
-		};
-		const redirectUrl = addQueryArgs(
-			{
-				connect_after_checkout: true,
-				from_site_slug: context.query.site,
-				admin_url: context.query.redirect_to.split( '?' )[ 0 ],
-			},
-			context.path.replace( /checkout\/[^?/]+\//, 'checkout/jetpack/' )
-		);
-
-		checkout( contextWithSiteSlug, next );
-
-		expect( pageSpy ).toHaveBeenCalledWith( redirectUrl );
-		expect( next ).toHaveBeenCalled();
 	} );
 } );
