+++ client/my-sites/controller.js
@@ -86,7 +86,6 @@ import { isSupportSession } from 'calypso/state/support/selectors';
 import { setSelectedSiteId, setAllSitesSelected } from 'calypso/state/ui/actions';
 import { setLayoutFocus } from 'calypso/state/ui/layout-focus/actions';
 import { getSelectedSite, getSelectedSiteId } from 'calypso/state/ui/selectors';
-import { isContextSourceMyJetpack } from './checkout/utils';
 
 /*
  * @FIXME Shorthand, but I might get rid of this.
@@ -538,13 +537,6 @@ export function siteSelection( context, next ) {
 		'1' === context.query?.unlinked && context.pathname.match( /^\/checkout\/[^/]+\/jetpack_/i );
 
 	const shouldRenderNoSites = ! context.section.enableNoSites && ! isUnlinkedCheckout;
-
-	// Immediately continue if the user is logged out and coming from My Jetpack
-	// so it gets picked up by the checkout middleware and redirects to logged-out checkout
-	if ( isContextSourceMyJetpack( context ) && ! currentUser ) {
-		next();
-		return;
-	}
 
 	// The user doesn't have any sites: render `NoSitesMessage`
 	if ( currentUser && currentUser.site_count === 0 && shouldRenderNoSites ) {
